+++ src/app/Repositories/ViewVariable/PetitionTypeViewModelVariable.php
@@ -4,7 +4,7 @@ use PCI\Models\AbstractBaseModel;
 use PCI\Models\MovementType;
 
 /**
- * Class NoteTypeViewModelVariable
+ * Class PetitionTypeViewModelVariable
  * El proposito de esta clase en relacion a las otras ViewVariables es que esta
  * necesita informacion especifica que esta siendo generada en ::init() aparte
  * de mi ignorancia en hacer esto.
@@ -13,7 +13,7 @@ use PCI\Models\MovementType;
  * @author  Alejandro Granadillo <slayerfat@gmail.com>
  * @link    https://github.com/slayerfat/sistemaPCI Repositorio en linea.
  */
-class NoteTypeViewModelVariable extends ViewModelVariable
+class PetitionTypeViewModelVariable extends ViewModelVariable
 {
 
     /**
@@ -26,7 +26,7 @@ class NoteTypeViewModelVariable extends ViewModelVariable
      */
     public function __construct(
         AbstractBaseModel $model,
-        $resource = 'noteTypes'
+        $resource = 'petitionTypes'
     ) {
         parent::__construct($model, $resource);
 
@@ -42,7 +42,7 @@ class NoteTypeViewModelVariable extends ViewModelVariable
     private function init()
     {
         $this->setParent(MovementType::class);
-        $this->setDestView('noteTypes.show');
+        $this->setDestView('petitionTypes.show');
         $this->setForeignKey('movement_type_id');
         $this->setParentTitle(trans('models.movementTypes.singular'));
     }